pyproject.toml
[project]
name = "audit-service"
version = "0.1.0"
description = "Audit system based on FastAPI and Cassandra (3-node cluster)"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "cassandra-driver>=3.29.2",
    "aiocassandra>=2.0.1",
    "pydantic>=2.9.0",
    "httpx>=0.27.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
